refactor: move page components to src/pages directory
