+++ src/components/componentsIndex.js
@@ -1,13 +1,13 @@
-export { default as ErrorBoundary } from './error-boundary/ErrorBoundary.jsx';
-export { default as ErrorBoundaryInRoutes } from './error-boundary/ErrorBoundaryInRoutes.jsx';
+export { default as ErrorBoundary } from '../pages/error-boundary/ErrorBoundary.jsx';
+export { default as ErrorBoundaryInRoutes } from '../pages/error-boundary/ErrorBoundaryInRoutes.jsx';
 export { default as ProtectedRoute } from './ProtectedRoute.jsx';
 export { default as Layout } from './layout/Layout.jsx';
-export { default as Login } from './login/Login.jsx';
+export { default as Login } from '../pages/login/Login.jsx';
 export { default as CloseEye } from './svg/CloseEye.jsx';
 export { default as OpenEye } from './svg/OpenEye.jsx';
 export { default as Spinner } from './svg/Spinner.jsx';
 export { default as CheckMark } from './svg/CheckMark.jsx';
-export { default as NotFound } from './not-found/NotFound.jsx';
-export { default as Signup } from './signup/Signup.jsx';
-export { default as EmailSent } from './email-sent/EmailSent.jsx';
-export { default as ResendVerificationEmail } from './resend-verification-email/ResendVerificationEmail.jsx';
+export { default as NotFound } from '../pages/not-found/NotFound.jsx';
+export { default as Signup } from '../pages/signup/Signup.jsx';
+export { default as EmailSent } from '../pages/email-sent/EmailSent.jsx';
+export { default as ResendVerificationEmail } from '../pages/resend-verification-email/ResendVerificationEmail.jsx';
